requirements.txt
asgiref==3.7.2
attrs==23.2.0
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
geographiclib==2.0
gunicorn==20.1.0
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
Markdown==3.5.2
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
sqlparse==0.4.4
uritemplate==4.1.1
whitenoise==6.4.0
